requirements.txt
django-cors-headers==3.10.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
django==3.1.12
djongo==1.3.6
pillow==8.4.0
pymongo==3.12.1

